Update favorite folders view
