--- Sources/Recycling/ListViews/FavoriteFolders/Model/FavoriteFoldersListViewModel.swift
@@ -1,19 +1,0 @@
-//
-//  Copyright © FINN.no AS, Inc. All rights reserved.
-//
-
-import Foundation
-
-public protocol FavoriteFoldersListViewModel {
-    var title: String { get }
-    var detail: String { get }
-    var accessibilityLabel: String { get }
-}
-
-public extension FavoriteFoldersListViewModel {
-    var accessibilityLabel: String {
-        var message = detail
-        message += ". " + title
-        return message
-    }
-}
